+++ .eslintrc.js
@@ -320,7 +320,7 @@ module.exports = {
 		"semi-style": "off",
 		"sort-imports": "off",
 		"sort-keys": "off",
-		"sort-vars": "error",
+		"sort-vars": "off",
 		"space-before-blocks": "error",
 		"space-before-function-paren": "off",
 		"space-in-parens": ["error", "never"],