+++ packages/host/src/export.ts
@@ -18,11 +18,10 @@ interface SimpleIconSpecification {
 	icon_size: number;
 	icon_mipmaps?: number;
 }
-type Color = { r: number, g: number, b: number, a: number };
 interface IconLayer {
 	icon: string;
 	icon_size: number;
-	tint?: Color,
+	tint?: lib.FactorioColor,
 	shift?: [number, number],
 	scale?: number,
 	icon_mipmaps?: number,
@@ -202,15 +201,9 @@ async function loadLayeredIcon(
 
 		iconLayer = iconLayer.clone();
 
-		let tint: Color;
+		let tint: lib.Color;
 		if (layer.tint) {
-			let divisor = (layer.tint.r > 1 || layer.tint.g > 1 || layer.tint.b > 1) ? 255 : 1;
-			tint = {
-				r: (layer.tint.r || 0) / divisor,
-				g: (layer.tint.g || 0) / divisor,
-				b: (layer.tint.b || 0) / divisor,
-				a: layer.tint.a || 1,
-			};
+			tint = lib.normalizeColor(layer.tint);
 		} else {
 			tint = { r: 1, b: 1, g: 1, a: 1 };
 		}
@@ -252,6 +245,18 @@ async function loadLayeredIcon(
 	return icon;
 }
 
+function fixIcons(item: ItemPrototype) {
+	const icons = item.icons;
+	if (typeof icons === "object" && !(icons instanceof Array) && icons !== null) {
+		// It's possible to specify icons as an object with arbitrary keys
+		// and the game will still accept it, cast the value to array if
+		// this is the case.
+		item = { ...item };
+		item.icons = Object.values(icons);
+	}
+	return item;
+}
+
 const itemTypes = new Set([
 	"item",
 	"ammo",
@@ -284,7 +289,8 @@ const itemTypes = new Set([
 function filterItems(prototypes: Prototypes): ItemPrototype[] {
 	return Object.entries(prototypes)
 		.filter(([type, _]) => itemTypes.has(type))
-		.flatMap(([_, typePrototypes]) => Object.values(typePrototypes)) as ItemPrototype[]
+		.flatMap(([_, typePrototypes]) => Object.values(typePrototypes) as ItemPrototype[])
+		.map(fixIcons)
 	;
 }